requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.128.5
h11==0.16.0
idna==3.11
psycopg==3.3.2
psycopg-binary==3.3.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
yt-dlp==2026.2.4
